+++ src/api/objects/account.ts
@@ -6,6 +6,7 @@ import { mastodon } from "masto";
 import { AccountLike, AccountNames } from "../../types";
 import { extractDomain } from "../../helpers";
 import { MastoApi } from "../api";
+import { replaceEmojiShortcodesWithImageTags } from "../../helpers";
 
 
 // Build a dict from the acct (e.g @user@server.com) to the Account object for easy lookup
@@ -48,4 +49,9 @@ export function webfingerURI(account: AccountLike): string {
     } else {
         return `${account.acct}@${extractServer(account)}`;
     }
+};
+
+
+export function accountNameWithEmojis(account: mastodon.v1.Account): string {
+    return replaceEmojiShortcodesWithImageTags(account.displayName, account.emojis || []);
 };