+++ src/index.ts
@@ -26,6 +26,7 @@ import TrendingLinksScorer from './scorer/feature/trending_links_scorer';
 import TrendingTagsScorer from "./scorer/feature/trending_tags_scorer";
 import TrendingTootScorer from "./scorer/feature/trending_toots_scorer";
 import VideoAttachmentScorer from "./scorer/feature/video_attachment_scorer";
+import { accountNameWithEmojis } from './api/objects/account';
 import { buildAccountNames } from "./api/objects/account";
 import { createRandomString, incrementCount } from "./helpers";
 import { DEFAULT_WEIGHTS, buildNewFilterSettings } from "./config";
@@ -422,4 +423,5 @@ export {
     TypeFilterName,
     WeightName,
     Weights,
+    accountNameWithEmojis,
 };